Split Volunteer component into VolunteerActivity and VolunteerHistory
